Split up PermissionGroupCommands
