+++ src/Permissions/PermissionGroupCommands.php
@@ -9,7 +9,6 @@
 
 namespace WildPHP\Core\Permissions;
 
-
 use WildPHP\Core\Channels\Channel;
 use WildPHP\Core\Channels\ChannelCollection;
 use WildPHP\Core\Commands\Command;
@@ -20,14 +19,11 @@ use WildPHP\Core\Commands\ParameterDefinitions;
 use WildPHP\Core\Commands\StringParameter;
 use WildPHP\Core\ComponentContainer;
 use WildPHP\Core\Connection\Queue;
-use WildPHP\Core\ContainerTrait;
 use WildPHP\Core\Modules\BaseModule;
 use WildPHP\Core\Users\User;
 
 class PermissionGroupCommands extends BaseModule
 {
-	use ContainerTrait;
-
 	/**
 	 * PermissionCommands constructor.
 	 *
@@ -87,86 +83,6 @@ class PermissionGroupCommands extends BaseModule
 			),
 			['rmgroup', 'rmg', 'removegroup', '-group', '-g']);
 
-		CommandHandler::fromContainer($container)->registerCommand('addmember',
-			new Command(
-				[$this, 'addmemberCommand'],
-				new ParameterDefinitions(2, 2, [
-					'groupName' => new ExistingPermissionGroupParameter($permissionGroupCollection),
-					'nickname' => new StringParameter()
-				]),
-				new CommandHelp([
-					'Add a member to a group in the permissions system. Usage: addmember [group name] [nickname]'
-				]),
-				'addmember'
-			),
-			['+member', '+m']);
-
-		CommandHandler::fromContainer($container)->registerCommand('delmember',
-			new Command(
-				[$this, 'delmemberCommand'],
-				new ParameterDefinitions(2, 2, [
-					'groupName' => new ExistingPermissionGroupParameter($permissionGroupCollection),
-					'nickname' => new StringParameter()
-				]),
-				new CommandHelp([
-					'Remove a member from a group in the permissions system. Usage: delmember [group name] [nickname]'
-				]),
-				'delmember'
-			),
-			['-member', '-m']);
-
-		CommandHandler::fromContainer($container)->registerCommand('allow',
-			new Command(
-				[$this, 'allowCommand'],
-				new ParameterDefinitions(2, 2, [
-					'groupName' => new ExistingPermissionGroupParameter($permissionGroupCollection),
-					'permission' => new StringParameter()
-				]),
-				new CommandHelp([
-					'Add a permission to a permission group. Usage: allow [group name] [permission]'
-				]),
-				'allow'
-			));
-
-		CommandHandler::fromContainer($container)->registerCommand('deny',
-			new Command(
-				[$this, 'denyCommand'],
-				new ParameterDefinitions(2, 2, [
-					'groupName' => new ExistingPermissionGroupParameter($permissionGroupCollection),
-					'permission' => new StringParameter()
-				]),
-				new CommandHelp([
-					'Remove a permission from a permission group. Usage: deny [group name] [permission]'
-				]),
-				'deny'
-			));
-
-		CommandHandler::fromContainer($container)->registerCommand('lsmembers',
-			new Command(
-				[$this, 'lsmembersCommand'],
-				new ParameterDefinitions(1, 1, [
-					'groupName' => new ExistingPermissionGroupParameter($permissionGroupCollection)
-				]),
-				new CommandHelp([
-					'List all members in a permission group. Usage: lsmembers [group name]'
-				]),
-				'lsmembers'
-			),
-			['lsm']);
-
-		CommandHandler::fromContainer($container)->registerCommand('lsperms',
-			new Command(
-				[$this, 'lspermsCommand'],
-				new ParameterDefinitions(1, 1, [
-					'groupName' => new ExistingPermissionGroupParameter($permissionGroupCollection)
-				]),
-				new CommandHelp([
-					'List all members in a permission group. Usage: lsperms [group name]'
-				]),
-				'lsperms'
-			),
-			['lsp']);
-
 		CommandHandler::fromContainer($container)->registerCommand('linkgroup',
 			new Command(
 				[$this, 'linkgroupCommand'],
@@ -211,198 +127,6 @@ class PermissionGroupCommands extends BaseModule
 			['gi']);
 
 		$this->setContainer($container);
-	}
-
-	/**
-	 * @param Channel $source
-	 * @param User $user
-	 * @param $args
-	 * @param ComponentContainer $container
-	 */
-	public function allowCommand(Channel $source, User $user, $args, ComponentContainer $container)
-	{
-		list ($groupName, $permission) = $args;
-
-		/** @var PermissionGroup|false $group */
-		$group = PermissionGroupCollection::fromContainer($container)
-			->offsetGet($groupName);
-
-		$checks = [
-			'This group does not exist.' => empty($group),
-			'This group is already allowed to do that.' => $group ? $group->getAllowedPermissions()->contains($permission) : false
-		];
-
-		if (!$this->doChecks($checks, $source, $user))
-			return;
-
-		$group->getAllowedPermissions()
-			->append($permission);
-		Queue::fromContainer($container)
-			->privmsg($source->getName(), $user->getNickname() . ': This group is now allowed the permission "' . $permission . '"');
-	}
-
-	/**
-	 * @param Channel $source
-	 * @param User $user
-	 * @param $args
-	 * @param ComponentContainer $container
-	 */
-	public function denyCommand(Channel $source, User $user, $args, ComponentContainer $container)
-	{
-		list ($groupName, $permission) = $args;
-
-		/** @var PermissionGroup|false $group */
-		$group = PermissionGroupCollection::fromContainer($container)
-			->offsetGet($groupName);
-
-		$checks = [
-			'This group does not exist.' => empty($group),
-			'This group is not allowed to do that.' => $group ? !$group->getAllowedPermissions()->contains($permission) : false
-		];
-
-		if (!$this->doChecks($checks, $source, $user))
-			return;
-
-		$group->getAllowedPermissions()
-			->removeAll($permission);
-		Queue::fromContainer($container)
-			->privmsg($source->getName(), $user->getNickname() . ': This group is now denied the permission "' . $permission . '"');
-	}
-
-	/**
-	 * @param Channel $source
-	 * @param User $user
-	 * @param $args
-	 * @param ComponentContainer $container
-	 */
-	public function lspermsCommand(Channel $source, User $user, $args, ComponentContainer $container)
-	{
-		$groupName = $args['groupName'];
-
-		$checks = [
-			'This group does not exist.' => !PermissionGroupCollection::fromContainer($container)
-				->offsetExists($groupName)
-		];
-
-		if (!$this->doChecks($checks, $source, $user))
-			return;
-
-		/** @var PermissionGroup|false $group */
-		$group = PermissionGroupCollection::fromContainer($container)
-			->offsetGet($groupName);
-
-		$perms = $group->getAllowedPermissions()
-			->values();
-		Queue::fromContainer($container)
-			->privmsg($source->getName(),
-				$user->getNickname() . ': The following permissions are set for this group: ' . implode(', ', $perms));
-	}
-
-	/**
-	 * @param Channel $source
-	 * @param User $user
-	 * @param $args
-	 * @param ComponentContainer $container
-	 */
-	public function lsmembersCommand(Channel $source, User $user, $args, ComponentContainer $container)
-	{
-		$groupName = $args['groupName'];
-
-		/** @var PermissionGroup|false $group */
-		$group = PermissionGroupCollection::fromContainer($container)
-			->offsetGet($groupName);
-
-		$checks = [
-			'This group does not exist.' => empty($group),
-			'This group may not contain members.' => $group ? $group->isModeGroup() : false
-		];
-
-		if (!$this->doChecks($checks, $source, $user))
-			return;
-
-		$members = $group->getUserCollection()
-			->values();
-		Queue::fromContainer($container)
-			->privmsg($source->getName(), sprintf('%s: The following members are in this group: %s',
-				$user->getNickname(),
-				implode(', ', $members)));
-	}
-
-	/**
-	 * @param Channel $source
-	 * @param User $user
-	 * @param $args
-	 * @param ComponentContainer $container
-	 */
-	public function addmemberCommand(Channel $source, User $user, $args, ComponentContainer $container)
-	{
-		list ($groupName, $nickname) = $args;
-
-		/** @var PermissionGroup|false $group */
-		$group = PermissionGroupCollection::fromContainer($container)
-			->offsetGet($groupName);
-
-		/** @var User $userToAdd */
-		$userToAdd = $source->getUserCollection()->findByNickname($nickname);
-
-		$checks = [
-			'This group does not exist.' => empty($group),
-			'This group may not contain members.' => $group ? $group->isModeGroup() : false,
-			'This user is not in my current database, in this channel, or is not logged in to services.' =>
-				empty($userToAdd) || empty($userToAdd->getIrcAccount()) || in_array($userToAdd->getIrcAccount(), ['*', '0'])
-		];
-
-		if (!$this->doChecks($checks, $source, $user))
-			return;
-
-		$group->getUserCollection()
-			->append($userToAdd->getIrcAccount());
-
-		Queue::fromContainer($container)
-			->privmsg($source->getName(),
-				sprintf('%s: User %s (identified by %s) has been added to the permission group "%s"',
-					$user->getNickname(),
-					$nickname,
-					$userToAdd->getIrcAccount(),
-					$groupName));
-	}
-
-	/**
-	 * @param Channel $source
-	 * @param User $user
-	 * @param $args
-	 * @param ComponentContainer $container
-	 */
-	public function delmemberCommand(Channel $source, User $user, $args, ComponentContainer $container)
-	{
-		list ($groupName, $nickname) = $args;
-
-		/** @var PermissionGroup|false $group */
-		$group = PermissionGroupCollection::fromContainer($container)
-			->offsetGet($groupName);
-
-		/** @var User $userToAdd */
-		$userToAdd = $source->getUserCollection()->findByNickname($nickname);
-
-		$checks = [
-			'This group does not exist.' => empty($group),
-			'This user is not in the group, in this channel, or not online.' => empty($userToAdd) || ($group && !$group->getUserCollection()
-						->contains($userToAdd->getIrcAccount()))
-		];
-
-		if (!$this->doChecks($checks, $source, $user))
-			return;
-
-		$group->getUserCollection()
-			->removeAll($userToAdd->getIrcAccount());
-
-		Queue::fromContainer($container)
-			->privmsg($source->getName(),
-				sprintf('%s: User %s (identified by %s) has been removed from the permission group "%s"',
-					$user->getNickname(),
-					$nickname,
-					$userToAdd->getIrcAccount(),
-					$groupName));
 	}
 
 	/**
@@ -620,29 +344,6 @@ class PermissionGroupCommands extends BaseModule
 		foreach ($lines as $line)
 			Queue::fromContainer($container)
 				->privmsg($source->getName(), $line);
-	}
-
-	/**
-	 * @param array $checks
-	 * @param Channel $source
-	 * @param User $user
-	 *
-	 * @return bool
-	 */
-	protected function doChecks(array $checks, Channel $source, User $user)
-	{
-		foreach ($checks as $string => $check)
-		{
-			if (!$check)
-				continue;
-
-			Queue::fromContainer($this->getContainer())
-				->privmsg($source->getName(), $user->getNickname() . ': ' . $string);
-
-			return false;
-		}
-
-		return true;
 	}
 
 	/**
