+++ src/bootstrap.php
@@ -147,6 +147,8 @@ function createNewInstance(\React\EventLoop\LoopInterface $loop, Configuration $
 		\WildPHP\Core\Channels\ChannelStateManager::class,
 		\WildPHP\Core\Commands\HelpCommand::class,
 		\WildPHP\Core\Permissions\PermissionGroupCommands::class,
+		\WildPHP\Core\Permissions\PermissionCommands::class,
+		\WildPHP\Core\Permissions\PermissionMembersCommands::class,
 		\WildPHP\Core\Management\ManagementCommands::class,
 		\WildPHP\Core\Users\BotStateManager::class,
 		\WildPHP\Core\Connection\NicknameHandler::class,
